feat: create new table for operator sensor status in foreman. remove exposure in teams table
